Update to project - Ray Casting and Line Render
Added script to left and right controllers. Added area and collision nodes to the each controller - linked them to signals when they enter other nodes with an "area" sub node. Added ray casting script to the left controller. Added the line render script from Canvas and attached them to the left controller. Added colors to to the line render.