Update to project
Added different color profiles for the left laser and boxes and for the right laser and boxes. Added new scene for the boxes to be spawned and for a start button to begin the game. Added the physical structure, transformed, and resize for the button and added coloring. to them.