Update to block spawner mechanic and right/left line render
Made the block object its own scene. Removed the block object from the spawner scene. Enabled shadow effect. Added a timer and marker node to the spawner scene. Connect the timer node to the spawner scene via timer out node. In the spawner scene, I added code to load and instantiate the block based on the timer out function. Fixed the left and right controller line render removal if a/x is pressed to be a vector3D instead of null.