Update to project (main and block scenes)
Installed the shatter plugin - will try to use this later. Added the music track - will implement this when ready. Changed the block node type from CSGBox3D to a MeshInstance3D to apply fracture effect. Added a collision to the block. Reorganized the block and collision to be child of a rigid body to apply physics. Added the block damage scene I created from Blender as a tscn. I removed the XR nodes to its own tscn called xr_player.