Update to Block Spawner and Blocks
Rearranged the block and spawner to be a specific spawner for spawning right/blue blocks and left/red blocks respectively. Added script to add random height for blocks to spawn. Adjusted the speed for blocks to spawn. Added the spawners to the main scene. Tested out the scene and confirm, blocks spawn. For now, the physics of when the blocks are hit and fall have been commented out.