Update to Project
Remove unused files (music track, start button). Added block damage sound source to README. Added recenter the user camera when menu button on the left controller is pressed. I did try adding reference nodes to the main scene for the camera position and orientation to match the reference but this did not work; using an XRServer class method to recenter the user.